envs p31

+++ apps/backend/src/controllers/envs/delete-env/delete-env.controller.ts
@@ -1,5 +1,6 @@
 import { Controller, Post } from '@nestjs/common';
 import { apiToBackend } from '~backend/barrels/api-to-backend';
+import { common } from '~backend/barrels/common';
 import { entities } from '~backend/barrels/entities';
 import { repositories } from '~backend/barrels/repositories';
 import { AttachUser, ValidateRequest } from '~backend/decorators/_index';
@@ -30,6 +31,12 @@ export class DeleteEnvController {
       memberId: user.user_id,
       projectId: projectId
     });
+
+    if (envId === common.PROJECT_ENV_PROD) {
+      throw new common.ServerError({
+        message: common.ErEnum.BACKEND_ENV_PROD_CAN_NOT_BE_DELETED
+      });
+    }
 
     await this.envsRepository.delete({
       project_id: projectId,
